% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{dataTableImportWidget}
\alias{dataTableImportWidget}
\title{Shiny UI widget to import a data table as data frame from file}
\usage{
dataTableImportWidget(id, label = "Comma or Tab Separated File")
}
\arguments{
\item{id}{the UI id for this widget}

\item{label}{the label of this widget}
}
\value{
A tagList that includes all the UI components
}
\description{
This widget provides UI for the user to select a local file
and specify options to parse the file to a data table. The preview panel will
update accordingly so users will be able to know if they are doing
things correctly.
}
