% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_preview.R
\name{dataTablePreviewWidget}
\alias{dataTablePreviewWidget}
\title{A data table preview widget}
\usage{
dataTablePreviewWidget(id, enableOptionToShowAllRows = TRUE)
}
\arguments{
\item{id}{The same ID as used in the matching call to \link{previewDataTable}}

\item{enableOptionToShowAllRows}{Whether or not to show a checkbox that toggles all rows all only the first few rows}
}
\value{
The widget
}
\description{
A data table preview widget
}
\examples{
# In Shiny UI logic, add this widget
dataTablePreviewWidget("previewTable")
# Then in your Shiny server logic, add following statements
dataTable <- reactive({ # generate your data frame here })
previewDataTable("previewTable", dataTable)
}
