% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{.dataTableImport}
\alias{.dataTableImport}
\title{Serve logic of file uploader UI}
\usage{
.dataTableImport(input, output, session, fileLocation = c("local",
  "server", "both"), serverRootFolders = c("."))
}
\arguments{
\item{input}{Shiny module inputs}

\item{session}{Shiny session}

\item{fileLocation}{specify from which location the file should be selected from}

\item{serverRootFolders}{the root folders that you want user to navigate. only used when \code{fileLocation} is specified as 'server'}

\item{ouput}{Shiny module outputs}
}
\value{
the imported data object as a data frame
}
\description{
The module server logic that reacts to user input.
}
