% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{fileImportWidget}
\alias{fileImportWidget}
\title{Shiny UI widget to import data from a file}
\usage{
fileImportWidget(id, label = NULL, dataType = c(C_DATA_TYPE_TABLE,
  C_DATA_TYPE_MATRIX), enableDataTypeSelection = TRUE)
}
\arguments{
\item{id}{ID of this UI component}

\item{label}{A descriptive label}

\item{dataType}{Data type to parse the file as}

\item{enableDataTypeSelection}{Whether to let the user specify which data type to import the file as}
}
\value{
The widget
}
\description{
This widget provides UI for the user to select a file
and specify options to parse and import the file. The preview panel will
update accordingly so users will be able to know if they are doing
things correctly.
}
