% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{fileSelectWidget}
\alias{fileSelectWidget}
\title{Shiny UI widget to select a file}
\usage{
fileSelectWidget(id)
}
\arguments{
\item{id}{the UI id for this widget}
}
\value{
A tagList that includes all the UI components
}
\description{
This widget provides UI for the user to select a file that is from local machine or on the server.
When user clicks the button, a modal dialog will show up for user
to navigate the file system.
}
