% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{importDataTable}
\alias{importDataTable}
\title{Shiny serve logic for \code{dataTableImportWidget}.}
\usage{
importDataTable(id, fileLocation = c("local", "server", "both"),
  serverRootFolders = c(wd = "."))
}
\arguments{
\item{id}{The same ID as used in the matching call to \code{dataTableImportWidget}}

\item{fileLocation}{specify from which location the file should be selected from}

\item{serverRootFolders}{the root folders that you want your users to navigate.
This parameter will only be used when \code{fileLocation} is specified as 'server' or 'both' and must be a named vector 
such as c("server-dir1" = "/path/on/server/1/", "server-dir2" = "/path/on/server/2/")}
}
\value{
the imported data object as a reactive data frame
}
\description{
Server logic that reacts to user input such as updating the preview table.
This function must be called within a Shiny server function
}
