% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{importFile}
\alias{importFile}
\title{Call this function in Shiny server logic to activate \link{fileImportWidget}.}
\usage{
importFile(id, fileLocation = c(C_FILE_LOCATION_BOTH,
  C_FILE_LOCATION_LOCAL, C_FILE_LOCATION_SERVER),
  serverRootDirectories = NULL)
}
\arguments{
\item{id}{The same ID as used in the matching call to \link{fileImportWidget}}

\item{fileLocation}{Specify from which location the file should be selected from}

\item{serverRootDirectories}{The root directories that your app users are allowed to navigate.
It must be a named vector such as \code{c("server-dir1" = "/path/on/server/1/", "server-dir2" = "/path/on/server/2/")}.
This parameter will only be used when \code{fileLocation} is specified as
\code{C_FILE_LOCATION_SERVER} or \code{C_FILE_LOCATION_BOTH}}
}
\value{
The imported data as the specified type
}
\description{
Call this function in Shiny server logic to activate \link{fileImportWidget}.
}
