% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_import.R
\name{internalDataTableImportUI}
\alias{internalDataTableImportUI}
\title{Internal implementation of table import widget UI}
\usage{
internalDataTableImportUI(id, label = "Tab or Comma Separated File")
}
\description{
Internal implementation of table import widget UI
}
