% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{internalFileImportUI}
\alias{internalFileImportUI}
\title{Internal implementation of file importing UI}
\usage{
internalFileImportUI(id, dataType = c(C_DATA_TYPE_TABLE,
  C_DATA_TYPE_MATRIX), enableDataTypeSelection = TRUE)
}
\description{
Internal implementation of file importing UI
}
