% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{internalServerFileSelect}
\alias{internalServerFileSelect}
\title{Serve logic of server file select module}
\usage{
internalServerFileSelect(input, output, session, serverRootDirectories)
}
\arguments{
\item{input}{Shiny module inputs}

\item{output}{Shiny module outputs}

\item{session}{Shiny session}

\item{serverRootDirectories}{the root folders that you would like to grant the users access and it must be a named vector.}
}
\value{
the selected file path as a reactive value
}
\description{
The module server logic that reacts to user input.
}
