% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_preview.R
\name{previewDataTable}
\alias{previewDataTable}
\title{Call this function in Shiny server logic to activate \link{dataTablePreviewWidget}}
\usage{
previewDataTable(id, dataTable)
}
\arguments{
\item{id}{The same ID as used in the matching call to \link{dataTablePreviewWidget}}

\item{dataTable}{A reactive expression that returns a data table}
}
\description{
Call this function in Shiny server logic to activate \link{dataTablePreviewWidget}
}
