% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{selectFile}
\alias{selectFile}
\title{Shiny serve logic for \code{fileSelectWidget}.}
\usage{
selectFile(id, fileLocation = c("local", "server", "both"))
}
\arguments{
\item{id}{The same ID as used in the matching call to \code{fileSelectWidget}}

\item{fileLocation}{a parameter indicates where you want ther users to navigate the files.}
}
\value{
the selected file path as a reactive value
}
\description{
Server logic that reacts to user input such as updating the current selected file path.
This function must be called within a Shiny server function
}
