% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_directory_select.R
\name{selectServerDirectory}
\alias{selectServerDirectory}
\title{Shiny serve logic for \code{serverDirectorySelectWidget}.}
\usage{
selectServerDirectory(id, serverRootDirectories)
}
\arguments{
\item{id}{The same ID as used in the matching call to \code{serverDirectorySelectWidget}}

\item{serverRootDirectories}{the root directories that your app users are allowed to navigate and it must be a named vector.}
}
\value{
the selected server directory path as a reactive value
}
\description{
Server logic that reacts to user input such as updating the current selected directory path.
This function must be called within a Shiny server function
}
