% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_file_select.R
\name{selectServerFile}
\alias{selectServerFile}
\title{Shiny serve logic for \code{serverFileSelectWidget}.}
\usage{
selectServerFile(id, serverRootDirectories)
}
\arguments{
\item{id}{The same ID as used in the matching call to \code{serverFileSelectWidget}}

\item{serverRootDirectories}{Root directories that your app users are allowed to navigate and it must be a named vector.}
}
\value{
The selected server file path as a reactive value
}
\description{
Server logic that reacts to user input such as updating the current selected file path.
This function must be called within a Shiny server function
}
