% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{selectServerFolder}
\alias{selectServerFolder}
\title{Shiny serve logic for \code{serverFolderSelectWidget}.}
\usage{
selectServerFolder(id, serverRootFolders)
}
\arguments{
\item{id}{The same ID as used in the matching call to \code{serverFolderSelectWidget}}

\item{serverRootFolders}{the root folders that you allow your app users to navigate under}
}
\value{
the selected server folder path as a reactive value
}
\description{
Server logic that reacts to user input such as updating the current selected folder path.
This function must be called within a Shiny server function
}
