% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_directory_select_impl.R
\name{serverDirectorySelectWidget}
\alias{serverDirectorySelectWidget}
\title{Shiny UI widget to select a directory on the server}
\usage{
serverDirectorySelectWidget(id, label = "Please select a folder",
  buttonLabel = "Browse...")
}
\arguments{
\item{id}{the UI id for this widget}

\item{label}{the label of this widget}

\item{buttonLabel}{the label of the button}
}
\value{
The server directory selection component
}
\description{
This widget provides UI for the user to select a server directory
When user clicks the button, a modal dialog will show up for user
to navigate the server file system.
}
