% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_file_select_impl.R
\name{serverFileSelectWidget}
\alias{serverFileSelectWidget}
\title{Shiny UI widget to select a file on the server}
\usage{
serverFileSelectWidget(id, label = "Please select a file",
  buttonLabel = "Browse...")
}
\arguments{
\item{id}{the UI id for this widget}

\item{label}{the label of this widget}

\item{buttonLabel}{the label of the button}
}
\value{
A tagList that includes all the UI components
}
\description{
This widget provides UI for the user to select a server file
When user clicks the button, a modal dialog will show up for user
to navigate the server file system.
}
