% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{serverFolderSelectWidget}
\alias{serverFolderSelectWidget}
\title{Shiny UI widget to select a folder on the server}
\usage{
serverFolderSelectWidget(id, label = "Please select a folder",
  buttonLabel = "Browse...")
}
\arguments{
\item{id}{the UI id for this widget}

\item{label}{the label of this widget}

\item{buttonLabel}{the label of the button}
}
\value{
A tagList that includes all the UI components
}
\description{
This widget provides UI for the user to select a server folder.
When user clicks the button, a modal dialog will show up for user
to navigate the server file system.
}
