% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_import.R
\name{dataTableImportWidget}
\alias{dataTableImportWidget}
\title{Shiny UI widget to import a data table as data frame from file}
\usage{
dataTableImportWidget(id, title = "Tab or Comma Separated File")
}
\arguments{
\item{id}{ID of this UI component}

\item{title}{Title of the popup dialog}
}
\value{
The widget
}
\description{
This widget provides UI for the user to select a file
and specify options to parse the file to a data table. The preview panel will
update accordingly so users will be able to know if they are doing
things correctly.
}
