% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select.R
\name{fileSelectWidget}
\alias{fileSelectWidget}
\title{A Shiny UI widget used to select a file}
\usage{
fileSelectWidget(id, label = NULL)
}
\arguments{
\item{id}{ID of this UI component}
}
\value{
Return the UI component
}
\description{
This module provides a UI component for the user to select a file that is from local file system or on the server.
When user clicks the button, a modal dialog will show up for user to navigate the file system.
}
