% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_select_impl.R
\name{selectFile}
\alias{selectFile}
\title{Select a file}
\usage{
selectFile(id, fileLocation = c(C_FILE_LOCATION_LOCAL,
  C_FILE_LOCATION_SERVER, C_FILE_LOCATION_BOTH),
  serverRootDirectories = NULL)
}
\arguments{
\item{id}{The same ID as used in the matching call to \link{fileSelectWidget}}

\item{fileLocation}{a parameter indicates where you want ther users to navigate the files.}

\item{serverRootDirectories}{the root directories that you allow your app users to navigate under}
}
\value{
the selected file path as a reactive value
}
\description{
Server logic that reacts to user input such as updating the current selected file path.
This function must be called within a Shiny server function
}
