% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_file_select.R
\name{serverFileSelectWidget}
\alias{serverFileSelectWidget}
\title{Shiny UI widget to select a file on the server}
\usage{
serverFileSelectWidget(id, title = "Please select a file",
  buttonLabel = "Browse...")
}
\arguments{
\item{id}{ID of this UI widget}

\item{title}{Title of the popup dialog}

\item{buttonLabel}{Label of the button}
}
\value{
The widget
}
\description{
This widget provides UI for the user to select a server file
When user clicks the button, a modal dialog will show up for user
to navigate the server file system.
}
